+++ src/test/java/io/parallec/core/main/tcp/sampleserver/TcpServerThread.java
@@ -1,5 +1,8 @@
 package io.parallec.core.main.tcp.sampleserver;
 
+import io.parallec.core.util.PcStringUtils;
+
+import java.io.IOException;
 import java.util.logging.Logger;
 
 public class TcpServerThread extends Thread {
@@ -11,11 +14,12 @@ public class TcpServerThread extends Thread {
     public TcpServerThread(boolean idle){
         this.idle = idle;
     }
+    private TcpEchoServer server = null;
     @Override
     public void run() {
 
         try {
-            TcpEchoServer server = new TcpEchoServer(idle);
+            server = new TcpEchoServer(idle);
             server.serve();
             while (!this.isShutdown()) {
                 ;
@@ -23,7 +27,7 @@ public class TcpServerThread extends Thread {
             server.stop();
             logger.info("TCP Server Stopped..");
         } catch (Exception e) {
-            System.err.println("Couldn't start TCP server:\n" + e);
+            logger.info("Couldn't start TCP server: " + PcStringUtils.printStackTrace(e));
         }
 
     }
@@ -35,6 +39,14 @@ public class TcpServerThread extends Thread {
     //interrupt is useful
     public void setShutdown(boolean shutdown) {
         this.shutdown = shutdown;
+        
+        if(shutdown){
+            try {
+                this.server.getServerSocket().close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
         this.interrupt();
     }